Task-1/requirements.txt
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.24.3
python-dotenv==1.0.0 